chore(STRK-225): extract shared vault-fixtures test helper
Codacy flagged the in-page vault-encryption chain as duplicated across two
core cloud-sync specs (PR #1306). `encryptVaultPayload` in attachments-cloud
and `encryptVaultFile` in item-price-history-cloud were near-identical: both
run the vaultRandomBytes/vaultDeriveKey/vaultEncrypt/serializeVaultFile chain
in-page to build a `salt|iv|iter|ciphertext` .stvault file under the composite
SYNC_KEY.

Extract a single `tests/playwright/helpers/vault-fixtures.js` exporting
`encryptVaultPayload(page, payload, syncKey)` and import it into both specs,
passing each spec's SYNC_KEY as an argument. Keeps the single-`page.evaluate`
shape (one round-trip, in-page iterations fallback).

Pure test refactor — no test cases added/removed (17 specs still green), no
coverage-map row needed, no behaviour change.
